Lint tweaks
