+++ packages/graphcap-ui/src/components/Fields/Input/index.stories.tsx
@@ -27,14 +27,14 @@ export const Password: Story = {
   },
 };
 
-export const Number: Story = {
+export const NumberInput: Story = {
   args: {
     type: "number",
     placeholder: "Number",
   },
 };
 
-export const Date: Story = {
+export const DateInput: Story = {
   args: {
     type: "date",
     placeholder: "Date",
