+++ packages/graphcap-ui/src/components/Fields/Select/Select.stories.tsx
@@ -275,19 +275,19 @@ export const IconsInOptions: Story = {
       <SelectContent>
         <SelectItem value="apple">
           <span className="flex items-center gap-2">
-            <span className="h-4 w-4 rounded-full bg-red-500" />
+            <span className="h-4 w-4 rounded-full bg-red-500" />{' '}
             Apple
           </span>
         </SelectItem>
         <SelectItem value="banana">
           <span className="flex items-center gap-2">
-            <span className="h-4 w-4 rounded-full bg-yellow-500" />
+            <span className="h-4 w-4 rounded-full bg-yellow-500" />{' '}
             Banana
           </span>
         </SelectItem>
         <SelectItem value="blueberry">
           <span className="flex items-center gap-2">
-            <span className="h-4 w-4 rounded-full bg-blue-500" />
+            <span className="h-4 w-4 rounded-full bg-blue-500" />{' '}
             Blueberry
           </span>
         </SelectItem>
