+++ packages/graphcap-ui/src/components/index.tsx
@@ -1,7 +1,8 @@
 export * from "./Button";
 export * from "./Button/LinkButton";
+export * from "./Fields/Input";
 export * from "./Fields/Text/Text";
-export * from "./Input";
 export * from "./Layout";
 export * from "./Notifications/Toaster";
 export * from "./Table";
+
